datepicker暴露出show和hide事件，picker滚动允许小数 (#1557)
* datetime插件暴露on-show和on-hide 事件

demo页（https://vux.li/demos/v2/#/component/datetime）在安卓上表现正常，在ios上有滚动穿透，不暴露事件，无法对背景进行操作（body{touch-action:none}）。

* 修改单次滚动距离，允许为小数

用了lib-flexible等缩放页面的情况下，极有可能出现小数

+++ src/components/picker/scroller.js
@@ -58,7 +58,7 @@ var Scroller = function (container, options) {
 
   self.__container.appendChild(component)
 
-  self.__itemHeight = parseInt(getComputedStyle(indicator, 'height'), 10)
+  self.__itemHeight = parseFloat(getComputedStyle(indicator, 'height'), 10)
 
   self.__callback = options.callback || function (top) {
     content.style.webkitTransform = 'translate3d(0, ' + (-top) + 'px, 0)'
@@ -71,7 +71,7 @@ var Scroller = function (container, options) {
   self.__setDimensions(component.clientHeight, content.offsetHeight)
 
   if (component.clientHeight === 0) {
-    self.__setDimensions(parseInt(getComputedStyle(component, 'height'), 10), 204)
+    self.__setDimensions(parseFloat(getComputedStyle(component, 'height'), 10), 204)
   }
   self.select(self.options.defaultValue, false)
